requirements.txt
subinitial.automation @ git+https://subinitial@bitbucket.org/subinitial/subinitial-automation.git
esptool
pybase62
blessed
pyserial
numpy